# go-user-auth-main/src/docs/swagger.yaml
basePath: /api
definitions:
...
  model.User:
    properties:
      _id:
        type: string
      email:
        type: string
      location:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      salt:
        type: string
      userId:
        type: integer
      username:
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  responses.UserResponse:
    properties:
      data:
        $ref: '#/definitions/model.User'
      error: {}
      message:
        type: string
      status:
        type: integer
    type: object
  responses.UsersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.User'
        type: array
      error: {}
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
  description: JWT Authentication header for all requests
  version: "3.0"
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Login
      tags:
      - Users
  /user/reset:
    post:
      consumes:
      - application/json
      description: ResetPassword
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "400":
          description: Failed to reset password
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: ResetPassword
      tags:
      - Users
  /v1/users:
    get:
      description: Retrieves all users.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UsersResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.UsersResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided details.
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new user
      tags:
      - Users
  /v1/users/{userid}:
    delete:
      description: Deletes a user by the given ID.
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a user by ID
      tags:
      - Users
    get:
      description: Retrieves a user by the given ID.
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates an existing user with the provided details.
      parameters:
      - description: User ID
        in: path
        name: userid
        required: true
        type: integer
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing user
      tags:
      - Users
swagger: "2.0"